Closes #19
Refactored the ecto type support. Now only `Interval.Support.EctoType`
does compile-time checks for Ecto and Postgres.

The module is _always_ defined, but functionality is conditionally defined.
The module hides this from the user, so now it is possible to
`use Interval.Support.EctoType` to include `Ecto.Type` support when it
is possible to generate.
